+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Not released yet
+
+- Automatically add `content-type: application/json` header if response is an
+  array. Previously this automatic header addition would only occur if JSON
+  body was an object at its root - array was not supported.
+
 ## 1.4.0
 
 IMPROVEMENTS